https://github.com/paypal/squbs/issues/310: Moved all streaming package members to parent package, except squbs-streamingpipeline.
