+++ squbs-admin/src/main/scala/org/squbs/admin/AdminSvc.scala
@@ -17,10 +17,9 @@ package org.squbs.admin
 
 import akka.http.scaladsl.model.Uri.Path
 import akka.http.scaladsl.model._
-import org.json4s.JsonAST.{JString, JObject}
+import org.json4s.JsonAST.{JObject, JString}
 import org.json4s.jackson.JsonMethods._
-import org.squbs.unicomplex.streaming.RouteDefinition
-import org.squbs.unicomplex.{ConfigUtil, WebContext}
+import org.squbs.unicomplex.{ConfigUtil, RouteDefinition, WebContext}
 import ConfigUtil._
 
 class AdminSvc extends RouteDefinition with WebContext {
