--- squbs-pattern/src/main/scala/org/squbs/pattern/spray/japi/ContentTypes.scala
@@ -1,26 +1,0 @@
-/*
- *  Copyright 2015 PayPal
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package org.squbs.pattern.spray.japi
-
-import spray.http.ContentTypes._
-
-object ContentTypes {
-
-  final val APPLICATION_JSON = `application/json`
-  final val APPLICATION_OCTET_STREAM = `application/octet-stream`
-  final val TEXT_PLAIN = `text/plain`
-  final val TEXT_PLAIN_UTF8 = `text/plain(UTF-8)`
-}
