--- squbs-pattern/src/test/java/org/squbs/pattern/spray/japi/MediaTypesTest.java
@@ -1,37 +1,0 @@
-/*
- *  Copyright 2015 PayPal
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-package org.squbs.pattern.spray.japi;
-
-import org.junit.Test;
-import spray.http.MediaType;
-
-import static org.junit.Assert.*;
-
-public class MediaTypesTest {
-
-    @Test
-    public void testByName() {
-        MediaType mediaType = MediaTypes.byName("application/json");
-        assertSame(MediaTypes.APPLICATION_JSON(), mediaType);
-    }
-
-    @Test
-    public void testCustomByName() {
-        MediaType m1 = MediaTypes.byName("application/foobar");
-        MediaType m2 = MediaTypes.byName("application/foobar");
-        assertSame(m1, m2);
-    }
-}
