+++ squbs-unicomplex/src/test/scala/org/squbs/dispatcher/ForkJoinConfiguratorSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.squbs.dispatcher.streaming
+package org.squbs.dispatcher
 
 import javax.management.ObjectName
