+++ squbs-unicomplex/src/test/scala/org/squbs/unicomplex/LocalPortListenerSpec.scala
@@ -25,7 +25,6 @@ import com.typesafe.config.ConfigFactory
 import org.scalatest.{BeforeAndAfterAll, FlatSpecLike, Matchers}
 import org.squbs.lifecycle.GracefulStop
 import org.squbs.unicomplex.Timeouts._
-import org.squbs.unicomplex.streaming.{LocalPortHeader, RouteDefinition}
 
 import scala.concurrent.Await
