+++ squbs-unicomplex/src/test/scala/org/squbs/unicomplex/MultiListenerSpec.scala
@@ -28,7 +28,6 @@ import com.typesafe.config.ConfigFactory
 import org.scalatest.{BeforeAndAfterAll, FlatSpecLike, Matchers}
 import org.squbs.lifecycle.GracefulStop
 import org.squbs.unicomplex.Timeouts._
-import org.squbs.unicomplex.streaming.RouteDefinition
 
 import scala.concurrent.Await
