+++ squbs-unicomplex/src/test/scala/org/squbs/unicomplex/pipeline/PipelineSpec.scala
@@ -29,11 +29,8 @@ import com.typesafe.config.ConfigFactory
 import org.scalatest.{BeforeAndAfterAll, FlatSpecLike, Matchers}
 import org.squbs.lifecycle.GracefulStop
 import org.squbs.pipeline.streaming._
-import org.squbs.unicomplex._
-import org.squbs.unicomplex.streaming.RouteDefinition
-import org.squbs.unicomplex.{JMX, PortBindings, Unicomplex, UnicomplexBoot}
-
 import org.squbs.unicomplex.Timeouts._
+import org.squbs.unicomplex._
 
 import scala.concurrent.Await
